requirements.txt
pytest
pytest-html
requests
requests-mock
pyyaml
openpyxl
pandas
pydantic[dotenv]
allure-pytest